+++ CHANGELOG.md
@@ -5,12 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 2025-10-06
 
 ### Added
 
 - Explicit option for `SkipClientVideo` in Setting (default: true)
-- Explicit option for `SuppressLoginNotification` in Settings (default: true)
+- Explicit option for `SuppressLoginNotification` in Settings (not implemented yet)
 - Message filter system using Regular Expressions
 
 ### Changed